requirements.txt
functions-framework==3.8.1
line-bot-sdk==3.13.0
python-dotenv==0.10.1
pyyaml==6.0.2
